internal/benchtest/Makefile: fix variable passing in release target
Previously, for the "release" target we shelled out to make so that we
could tee the output to a file, which makes passing any variables a pain
(since we need to pass all of them when invoking the make child
process). This commit changes the Makefile so that the existing targets
can conditionally tee to a file - thus we don't have to shell out.