strcase/bytcase: don't fold LATIN SMALL LETTER DOTLESS I
Change code to not fold 'ı' (LATIN SMALL LETTER DOTLESS I). The issue
here is that 'ı' and 'İ' have an upper/lower case, but no folds and we
were incorrectly using the upper case form of 'ı' in comparisons.

TODO: change the generation logic to exclude 'ı' and 'İ' from the
upper/lower case table.